# frontend/openapi.yml
openapi: 3.0.3
info:
  title: API PENT
  version: 1.0.0
  description: Façade HTTP typée vers les règles métier Odoo PENT.
paths:
...
  /api/v1/proposals/{id}/:
    get:
      operationId: proposal_retrieve
      description: Authenticate via the relayed Odoo session and normalize RPC failures.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      tags:
        - proposals
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ""
    patch:
      operationId: proposals_partial_update
      description: Authenticate via the relayed Odoo session and normalize RPC failures.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      tags:
        - proposals
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchedProposalUpdate"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/PatchedProposalUpdate"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PatchedProposalUpdate"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ""
...
  /api/v1/team/{id}/:
    get:
      operationId: team_employee
      description: Authenticate via the relayed Odoo session and normalize RPC failures.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      tags:
        - team
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ""
...
components:
  schemas:
...
    PatchedProposalUpdate:
      type: object
      properties:
        start_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        estimated_work_days:
          type: string
          format: decimal
          pattern: ^-?\d{0,6}(?:\.\d{0,1})?$
        title:
          type: string
          maxLength: 180
        description:
          type: string
        action_id:
          type: integer
          minimum: 1
          nullable: true
        calendar_id:
          type: integer
          minimum: 1
        revision:
          type: integer
          minimum: 1
...